refactor: dead code cleanup — remove shims, fix dead tests checker
Dead src/ code removed:
- src/worktree/dispatcher.ts — dead DispatchResult interface, zero callers
- src/verification/parser.ts — deprecated re-export shim; migrated 6 src/
  callers and 2 test files to import directly from src/test-runners

Test fixes:
- test/unit/prompts/loader.test.ts — fix wrong import path (src/prompts/types
  → src/prompts/core/types), remove stale config keys (defaultAgent,
  dangerouslySkipPermissions, analyze) rejected by schema migrations
- test/unit/utils/test-output-parser.test.ts + rectification.test.ts —
  migrate from deprecated shim to src/test-runners

check-dead-tests.ts improvements:
- Remove "worktree" and "dispatcher" from REMOVED_FEATURES (both still active,
  caused 16 false positives out of 22 reported)
- Fix import scanning to only check before first describe/test/it block,
  eliminating 7 false positives from fixture strings in template literals
- Update checker's own test suite to use current REMOVED_FEATURES entries

Checker now reports 0 files with issues (was 22). All tests pass.
