+++ docs/dead-tests-report.md
@@ -1,0 +1,5 @@
+# Dead Tests Report
+
+Generated: 2026-05-21T08:04:09.818Z
+
+No dead tests detected. All test files are healthy!
