+++ src/verification/index.ts
@@ -7,7 +7,6 @@
 
 export * from "./types";
 export * from "./executor";
-export * from "./parser";
 export * from "./runners";
 export * from "./rectification";
 export * from "./shared-rectification-loop";
