+++ src/verification/strategies/regression.ts
@@ -10,9 +10,9 @@
  */
 
 import { getSafeLogger } from "../../logger";
+import { parseTestOutput } from "../../test-runners";
 import type { IVerificationStrategy, VerifyContext, VerifyResult } from "../orchestrator-types";
 import { makeFailResult, makePassResult, makeSkippedResult } from "../orchestrator-types";
-import { parseTestOutput } from "../parser";
 import { fullSuite } from "../runners";
 
 export class RegressionStrategy implements IVerificationStrategy {
