+++ src/verification/strategies/scoped.ts
@@ -11,10 +11,10 @@
  */
 
 import { getLogger } from "../../logger";
+import { parseTestOutput } from "../../test-runners";
 import { DEFAULT_TEST_FILE_PATTERNS, globsToTestRegex } from "../../test-runners/conventions";
 import type { IVerificationStrategy, StructuredTestFailure, VerifyContext, VerifyResult } from "../orchestrator-types";
 import { makeFailResult, makePassResult, makeSkippedResult } from "../orchestrator-types";
-import { parseTestOutput } from "../parser";
 import { regression } from "../runners";
 import { _smartRunnerDeps } from "../smart-runner";
