--- src/worktree/dispatcher.ts
@@ -1,6 +1,0 @@
-export interface DispatchResult {
-  storyId: string;
-  success: boolean;
-  worktreePath: string;
-  error?: string;
-}
